Add web and desktop sections to downloads page
