+++ src/app/downloads/downloads.module.css
@@ -1,0 +1,20 @@
+.cdn {
+  font-size: var(--mantine-font-size-md);
+}
+
+.download {
+  background-color: var(--ruffle-orange);
+
+  * {
+    color: #3c1518;
+  }
+
+  &[disabled] {
+    background-color: var(--ruffle-blue-7);
+    * {
+      color: var(--mantine-color-white);
+      font-weight: lighter;
+      text-decoration: line-through;
+    }
+  }
+}
