+++ src/app/downloads/nightlies.tsx
@@ -17,91 +17,12 @@ import React from "react";
 import Link from "next/link";
 import classes from "./nightlies.module.css";
 import {
-  IconBrandApple,
-  IconBrandChrome,
-  IconBrandDebian,
-  IconBrandFirefox,
-  IconBrandJavascript,
-  IconBrandSafari,
-  IconBrandWindows,
-  TablerIconsProps,
-} from "@tabler/icons-react";
-
-export interface GithubRelease {
-  id: number;
-  name: string;
-  prerelease: boolean;
-  downloads: ReleaseDownloads;
-  url: string;
-}
-
-export interface ReleaseDownloads {
-  windows_64?: string;
-  windows_32?: string;
-  macos?: string;
-  linux?: string;
-
-  firefox?: string;
-  chrome?: string;
-  safari?: string;
-
-  web?: string;
-}
-
-interface DownloadLink {
-  key: keyof ReleaseDownloads;
-  name: string;
-  icon: (props: TablerIconsProps) => React.JSX.Element;
-}
-
-const desktopLinks: DownloadLink[] = [
-  {
-    key: "windows_64",
-    name: "Windows (64-bit)",
-    icon: IconBrandWindows,
-  },
-  {
-    key: "windows_32",
-    name: "Windows (32-bit)",
-    icon: IconBrandWindows,
-  },
-  {
-    key: "macos",
-    name: "macOS",
-    icon: IconBrandApple,
-  },
-  {
-    key: "linux",
-    name: "Linux",
-    icon: IconBrandDebian,
-  },
-];
-
-const extensionLinks: DownloadLink[] = [
-  {
-    key: "chrome",
-    name: "Chrome / Edge",
-    icon: IconBrandChrome,
-  },
-  {
-    key: "firefox",
-    name: "Firefox",
-    icon: IconBrandFirefox,
-  },
-  {
-    key: "safari",
-    name: "Safari",
-    icon: IconBrandSafari,
-  },
-];
-
-const webLinks: DownloadLink[] = [
-  {
-    key: "web",
-    name: "Self Hosted",
-    icon: IconBrandJavascript,
-  },
-];
+  desktopLinks,
+  DownloadLink,
+  extensionLinks,
+  GithubRelease,
+  webLinks,
+} from "@/app/downloads/config";
 
 function DownloadLink({
   link,
@@ -192,7 +113,7 @@ function NightlyCompactBox(release: GithubRelease) {
 
 export function NightlyList({ nightlies }: { nightlies: GithubRelease[] }) {
   return (
-    <>
+    <Stack>
       <Title>Latest Nightly Releases</Title>
       <Text>
         If none of the above are suitable for you, you can manually download the
@@ -226,6 +147,6 @@ export function NightlyList({ nightlies }: { nightlies: GithubRelease[] }) {
           <NightlyCompactBox key={nightly.id} {...nightly} />
         ))}
       </Stack>
-    </>
+    </Stack>
   );
 }
